+++ spec/spec_helper.rb
@@ -8,6 +8,53 @@ require_relative "../lib/sidekiq-encrypted_args"
 RSpec.configure do |config|
   config.warnings = true
   config.order = :random
+
+  config.around(:each) do |example|
+    if example.metadata[:no_warn]
+      save_stderr = $stderr
+      begin
+        $stderr = StringIO.new
+        example.run
+      ensure
+        $stderr = save_stderr
+      end
+    else
+      example.run
+    end
+  end
+end
+
+# Reset all middleware for nested context and then restore.
+#
+# @note Middleware args are not preserved
+def with_empty_middleware
+  # Save the middleware context
+  server_middleware = Sidekiq.server_middleware.entries.map(&:klass)
+  client_middleware = Sidekiq.client_middleware.entries.map(&:klass)
+  Sidekiq.server_middleware.clear
+  Sidekiq.client_middleware.clear
+
+  yield
+
+  # Clear anything added and restore all previously registered middleware
+  Sidekiq.server_middleware.clear
+  Sidekiq.client_middleware.clear
+  server_middleware.each { |m| Sidekiq.server_middleware.add(m) }
+  client_middleware.each { |m| Sidekiq.client_middleware.add(m) }
+end
+
+def as_sidekiq_server!
+  allow(Sidekiq).to receive(:server?).and_return true
+end
+
+def as_sidekiq_client!
+  allow(Sidekiq).to receive(:server?).and_return false
+end
+
+class EmptyMiddleware
+  def call(*args)
+    yield
+  end
 end
 
 class RegularWorker
@@ -39,17 +86,17 @@ module Super
   class SecretWorker
     include Sidekiq::Worker
 
-    sidekiq_options encrypted_args: [false, false, true]
+    sidekiq_options encrypted_args: "arg_3"
 
     def perform(arg_1, arg_2, arg_3)
     end
   end
 end
 
-class ArrayOptionSecretWorker
+class ArrayIndexSecretWorker
   include Sidekiq::Worker
 
-  sidekiq_options encrypted_args: [false, true]
+  sidekiq_options encrypted_args: [1]
 
   def perform(arg_1, arg_2, arg_3)
   end
@@ -59,6 +106,15 @@ class HashOptionSecretWorker
   include Sidekiq::Worker
 
   sidekiq_options encrypted_args: {1 => true}
+
+  def perform(arg_1, arg_2, arg_3)
+  end
+end
+
+class ArrayOptionSecretWorker
+  include Sidekiq::Worker
+
+  sidekiq_options encrypted_args: [false, true]
 
   def perform(arg_1, arg_2, arg_3)
   end
@@ -79,5 +135,15 @@ class NamedHashOptionSecretWorker
   sidekiq_options encrypted_args: {arg_2: true, arg_1: false}
 
   def perform(arg_1, arg_2, arg_3)
+  end
+end
+
+class ComplexRubyType
+  def initialize(attributes)
+    @attributes = attributes
+  end
+
+  def to_json
+    @attributes.to_json
   end
 end