+++ CHANGE_LOG.md
@@ -1,4 +1,14 @@
-# Change log
+# Change Log
+
+## 1.1.0
+
+* Use `to_json` if it is defined when serializing encrypted args to JSON.
+* Add client middleware to the server default configuration. This ensures that arguments will be encrypted if a worker enqueues a job with encrypted arguments.
+* Client middleware now reads sidekiq options from the job hash instead of from the worker class so that the list of encrypted arguments is always in sync on the job payload.
+* Don't blow up if class name that is not defined is passed to client middleware.
+* Added additional option to specify encrypted args with array of argument indexes.
+* Deprecated setting encrypted args as hash or array of booleans.
+* Client middleware is prepended while server middleware is appended.
 
 ## 1.0.2